Utils exports for path
